requirements.txt
pandas==2.2.1
numpy==1.26.4
tensorflow-macos==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
scikit-learn==1.4.1.post1
flask==3.0.2
pytest==8.0.2
flask-cors==4.0.0